% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_census_data.R
\name{explore_census_data}
\alias{explore_census_data}
\title{Launch the Census ACS Variable Explorer}
\usage{
explore_census_data()
}
\description{
This function launches an interactive Shiny application to explore,
filter, and download Census ACS variables.
}
\examples{
\dontrun{
ACSloadR::explore_census_data()
}
}
