% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-acs-lmi.R
\name{get_acs_commuting_characteristics}
\alias{get_acs_commuting_characteristics}
\title{Get tidy ACS commuting characteristics data}
\usage{
get_acs_commuting_characteristics(
  year,
  survey = c("acs5", "acs1"),
  geography,
  ...,
  cache_table = TRUE
)
}
\arguments{
\item{year}{Numeric ACS vintage.}

\item{survey}{Either `"acs5"` or `"acs1"`.}

\item{geography}{A geography accepted by [tidycensus::get_acs()].}

\item{...}{Additional geography and geometry arguments forwarded to
[tidycensus::get_acs()]. Managed arguments such as `variables`, `table`,
and `output` cannot be supplied here.}

\item{cache_table}{Logical; cache Census metadata and downloaded tables.}
}
\value{
An `acs_lmi_bundle` containing mode cross-tabulation tibbles for residence and workplace geographies.
}
\description{
Downloads detailed transportation mode cross-tabulations by socio-economic characteristics for both residence and workplace geographies:
age (B08101/C08101, B08501/C08501), median age (B08103, B08503), earnings (B08119/C08119, B08519/C08519),
median earnings (B08121, B08521), poverty status (B08122/C08122, B08522/C08522), occupation (B08124/C08124, B08524/C08524),
industry (B08126/C08126, B08526/C08526), class of worker (B08128/C08128, B08528/C08528), vehicles available (B08141/C08141, B08541/C08541),
and racial iterations (B08105A-I, B08505A-I).
}
\examples{
\dontrun{
comm_char <- get_acs_commuting_characteristics(2024, "acs5", "state", state = "MA")
comm_char$commuting_earnings_residence
comm_char$commuting_vehicles_residence
}
}
