% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-acs-lmi.R
\name{get_acs_migration_current}
\alias{get_acs_migration_current}
\title{Get tidy ACS geographical mobility data by current residence}
\usage{
get_acs_migration_current(
  year,
  survey = c("acs5", "acs1"),
  geography,
  ...,
  cache_table = TRUE
)
}
\arguments{
\item{year}{Numeric ACS vintage.}

\item{survey}{Either `"acs5"` or `"acs1"`.}

\item{geography}{A geography accepted by [tidycensus::get_acs()].}

\item{...}{Additional geography and geometry arguments forwarded to
[tidycensus::get_acs()]. Managed arguments such as `variables`, `table`,
and `output` cannot be supplied here.}

\item{cache_table}{Logical; cache Census metadata and downloaded tables.}
}
\value{
An `acs_lmi_bundle` containing current residence geographical mobility tibbles.
}
\description{
Downloads detailed migration and geographical mobility tables for population in current residence
covering age (B07001/C07001), median age (B07002), sex (B07003), race (B07004A-I), citizenship (B07007),
marital status (B07008/C07008), education (B07009), income (B07010), median income (B07011), poverty (B07012),
tenure (B07013), region-to-region movers (B07101), Metro MSA (B07201/C07201), Micro MSA (B07202), Non-metro (B07203),
and State/County/Place level (B07204/C07204).
}
\examples{
\dontrun{
mig_curr <- get_acs_migration_current(2024, "acs5", "state", state = "MA")
mig_curr$migration_age
mig_curr$migration_income
}
}
