% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-acs-lmi.R
\name{get_acs_migration_prior}
\alias{get_acs_migration_prior}
\title{Get tidy ACS geographical mobility data by residence 1 year ago}
\usage{
get_acs_migration_prior(
  year,
  survey = c("acs5", "acs1"),
  geography,
  ...,
  cache_table = TRUE
)
}
\arguments{
\item{year}{Numeric ACS vintage.}

\item{survey}{Either `"acs5"` or `"acs1"`.}

\item{geography}{A geography accepted by [tidycensus::get_acs()].}

\item{...}{Additional geography and geometry arguments forwarded to
[tidycensus::get_acs()]. Managed arguments such as `variables`, `table`,
and `output` cannot be supplied here.}

\item{cache_table}{Logical; cache Census metadata and downloaded tables.}
}
\value{
An `acs_lmi_bundle` containing residence 1 year ago geographical mobility tibbles.
}
\description{
Downloads detailed migration and geographical mobility tables for population by residence 1 year ago
covering age (B07401/C07401), median age (B07402), sex (B07403), race (B07404A-I), citizenship (B07407),
marital status (B07408), education (B07409), income (B07410), median income (B07411), poverty (B07412),
and tenure (B07413).
}
\examples{
\dontrun{
mig_prior <- get_acs_migration_prior(2024, "acs5", "state", state = "MA")
mig_prior$migration_age
mig_prior$migration_income
}
}
