% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-acs-lmi.R
\name{get_acs_place_of_work}
\alias{get_acs_place_of_work}
\title{Get tidy ACS place of work geography data}
\usage{
get_acs_place_of_work(
  year,
  survey = c("acs5", "acs1"),
  geography,
  ...,
  cache_table = TRUE
)
}
\arguments{
\item{year}{Numeric ACS vintage.}

\item{survey}{Either `"acs5"` or `"acs1"`.}

\item{geography}{A geography accepted by [tidycensus::get_acs()].}

\item{...}{Additional geography and geometry arguments forwarded to
[tidycensus::get_acs()]. Managed arguments such as `variables`, `table`,
and `output` cannot be supplied here.}

\item{cache_table}{Logical; cache Census metadata and downloaded tables.}
}
\value{
An `acs_lmi_bundle` containing place of work location tibbles.
}
\description{
Downloads detailed place of work tables covering state & county level (B08007), place level (B08008),
MCD level (B08009), Metro MSA level (B08016/C08016), Micro MSA level (B08017), Non-metro level (B08018),
and total workplace worker population (B08604).
}
\examples{
\dontrun{
pow <- get_acs_place_of_work(2024, "acs5", "state", state = "MA")
pow$place_of_work_county
pow$workplace_worker_pop
}
}
