% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle.R
\name{new_acs_lmi_bundle}
\alias{new_acs_lmi_bundle}
\title{An ACS LMI data bundle}
\usage{
new_acs_lmi_bundle(components, topic, year, survey)
}
\arguments{
\item{components}{A named list of data frames.}

\item{topic}{Topic name used when printing the bundle.}

\item{year}{ACS vintage.}

\item{survey}{ACS survey dataset.}
}
\value{
An object of class `acs_lmi_bundle`.
}
\description{
Topic helpers return a named bundle of analysis-ready tibbles. Components
remain separate when their ACS source tables have different structures.
}
\keyword{internal}
