% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_select.R
\name{region_select}
\alias{region_select}
\title{Launch the Geographic Region Select Shiny App}
\usage{
region_select()
}
\description{
This function launches an interactive Shiny application that allows users to
select geographic regions (counties, tracts, ZCTAs, etc.) on a map and exports
the selected regions as R code for use in data analysis.
}
\examples{
\dontrun{
run_region_select()
}
}
